+++ xrpl/queries/account/info.go
@@ -4,6 +4,7 @@ import (
 	"github.com/Peersyst/xrpl-go/xrpl/ledger-entry-types"
 	accounttypes "github.com/Peersyst/xrpl-go/xrpl/queries/account/types"
 	"github.com/Peersyst/xrpl-go/xrpl/queries/common"
+	"github.com/Peersyst/xrpl-go/xrpl/queries/version"
 	"github.com/Peersyst/xrpl-go/xrpl/transaction/types"
 )
 
@@ -15,6 +16,7 @@ import (
 // activity, and its XRP balance. All information retrieved is relative to a
 // particular version of the ledger.
 type InfoRequest struct {
+	common.BaseRequest
 	Account     types.Address          `json:"account"`
 	LedgerIndex common.LedgerSpecifier `json:"ledger_index,omitempty"`
 	LedgerHash  common.LedgerHash      `json:"ledger_hash,omitempty"`
@@ -25,6 +27,10 @@ type InfoRequest struct {
 
 func (*InfoRequest) Method() string {
 	return "account_info"
+}
+
+func (*InfoRequest) APIVersion() int {
+	return version.RippledAPIV2
 }
 
 // TODO: Implement (V2)