+++ xrpl/queries/path/path_find.go
@@ -1,7 +1,9 @@
 package path
 
 import (
+	"github.com/Peersyst/xrpl-go/xrpl/queries/common"
 	pathtypes "github.com/Peersyst/xrpl-go/xrpl/queries/path/types"
+	"github.com/Peersyst/xrpl-go/xrpl/queries/version"
 	"github.com/Peersyst/xrpl-go/xrpl/transaction"
 	"github.com/Peersyst/xrpl-go/xrpl/transaction/types"
 )
@@ -20,6 +22,7 @@ const (
 
 // Start sending pathfinding information.
 type FindCreateRequest struct {
+	common.BaseRequest
 	Subcommand         SubCommand             `json:"subcommand"`
 	SourceAccount      types.Address          `json:"source_account,omitempty"`
 	DestinationAccount types.Address          `json:"destination_account,omitempty"`
@@ -30,6 +33,10 @@ type FindCreateRequest struct {
 
 func (*FindCreateRequest) Method() string {
 	return "path_find"
+}
+
+func (*FindCreateRequest) APIVersion() int {
+	return version.RippledAPIV1
 }
 
 // TODO: Implement V2
@@ -43,11 +50,16 @@ func (*FindCreateRequest) Validate() error {
 
 // Stop sending pathfinding information.
 type FindCloseRequest struct {
+	common.BaseRequest
 	Subcommand SubCommand `json:"subcommand"`
 }
 
 func (*FindCloseRequest) Method() string {
 	return "path_find"
+}
+
+func (*FindCloseRequest) APIVersion() int {
+	return version.RippledAPIV2
 }
 
 // TODO: Implement V2
@@ -61,11 +73,16 @@ func (*FindCloseRequest) Validate() error {
 
 // Get the information of the currently-open pathfinding request.
 type FindStatusRequest struct {
+	common.BaseRequest
 	Subcommand SubCommand `json:"subcommand"`
 }
 
 func (*FindStatusRequest) Method() string {
 	return "path_find"
+}
+
+func (*FindStatusRequest) APIVersion() int {
+	return version.RippledAPIV2
 }
 
 // TODO: Implement V2