+++ xrpl/queries/server/server_info.go
@@ -1,7 +1,9 @@
 package server
 
 import (
+	"github.com/Peersyst/xrpl-go/xrpl/queries/common"
 	servertypes "github.com/Peersyst/xrpl-go/xrpl/queries/server/types"
+	"github.com/Peersyst/xrpl-go/xrpl/queries/version"
 )
 
 // ############################################################################
@@ -11,10 +13,15 @@ import (
 // The server_info command asks the server for a human-readable version of
 // various information about the rippled server being queried.
 type InfoRequest struct {
+	common.BaseRequest
 }
 
 func (*InfoRequest) Method() string {
 	return "server_info"
+}
+
+func (*InfoRequest) APIVersion() int {
+	return version.RippledAPIV2
 }
 
 // TODO: Implement V2