+++ xrpl/queries/utility/random.go
@@ -1,6 +1,10 @@
 package utility
 
-import "github.com/Peersyst/xrpl-go/xrpl/transaction/types"
+import (
+	"github.com/Peersyst/xrpl-go/xrpl/queries/common"
+	"github.com/Peersyst/xrpl-go/xrpl/queries/version"
+	"github.com/Peersyst/xrpl-go/xrpl/transaction/types"
+)
 
 // ############################################################################
 // Request
@@ -8,10 +12,16 @@ import "github.com/Peersyst/xrpl-go/xrpl/transaction/types"
 
 // The random command provides a random number to be used as a source of
 // entropy for random number generation by clients.
-type RandomRequest struct{}
+type RandomRequest struct {
+	common.BaseRequest
+}
 
 func (*RandomRequest) Method() string {
 	return "random"
+}
+
+func (*RandomRequest) APIVersion() int {
+	return version.RippledAPIV2
 }
 
 func (*RandomRequest) Validate() error {